Adding in better logic to detect the poor snake's death
We now end the game when the snake has hit a wall or itself. :(
There is an artifact when we hit the negative walls (left and top) which
may be associated with under-running the usize on the snake's positions.
I think I'll convert the body to be an i32 as this will remove quite a
lot of 'as usize'... calls.